testing: add simple testing support
New script 'share/goa/lib/test.tcl'

The script looks for the projects defined in 'share/goa/lib/test.list'
(format: <test>[:timeout]) in the 'examples' folder of this project.  If
the project contains a "test/artifact" file, the script executes the
project and compares the output to the first line in the 'test/artifact'
file. In case the line matches the output it is evaluated as success.

issue genodelabs/goa#43