rust: use 'compat-libc' api/so
Link against 'compat-libc' api that declares 'stat@FBSD_1.0' and
'fstat@FBSD_1.0'. 'compat-libc.lib.so' implements these functions.
Unfortunately, the library has to be before 'libc.lib.so' during linkage
and has to be first in the NEEDED section as well. So 'stat' is found
before 'Libc::stat'. This breaks 'fork' on Genode.